# 10x faster Atom class