--- a/Framework/Doxygen/Mantid_template.doxyfile
+++ b/Framework/Doxygen/Mantid_template.doxyfile
@@ -158,7 +158,7 @@ EXCLUDE                = @CMAKE_CURRENT_SOURCE_DIR@/../ICat/src/GSoapGenerated \
                          @CMAKE_CURRENT_SOURCE_DIR@/../Kernel/src/ANN_complete.cpp \
                          @CMAKE_CURRENT_SOURCE_DIR@/../../qt/widgets/common/inc/MantidQtWidgets/Common/QtPropertyBrowser \
                          @CMAKE_CURRENT_SOURCE_DIR@/../../qt/widgets/common/src/QtPropertyBrowser \
-                         @CMAKE_CURRENT_SOURCE_DIR@/../../qt/paraview_ext/PVPlugins
+                         @CMAKE_CURRENT_SOURCE_DIR@/../../qt/paraview_ext/PVPlugins \
                          @CMAKE_CURRENT_SOURCE_DIR@/../Kernel/inc/MantidKernel/span.hpp
 
 EXCLUDE_SYMLINKS       = NO