--- a/Framework/Kernel/CMakeLists.txt
+++ b/Framework/Kernel/CMakeLists.txt
@@ -490,7 +490,7 @@ target_include_directories(Kernel SYSTEM
                            PUBLIC ${Boost_INCLUDE_DIRS} ${POCO_INCLUDE_DIRS}
                                   ${JSONCPP_INCLUDE_DIR}
                            PRIVATE ${NEXUS_INCLUDE_DIR} ${GSL_INCLUDE_DIR}
-                                   ${OPENSSL_INCLUDE_DIR})
+                                   ${OPENSSL_INCLUDE_DIR} ${SPAN_INCLUDE_DIR})
 
 if(OSX_VERSION VERSION_GREATER 10.8)
   set_target_properties(Kernel PROPERTIES INSTALL_RPATH "@loader_path/../MacOS")