--- a/Framework/Kernel/inc/MantidKernel/Atom.h
+++ b/Framework/Kernel/inc/MantidKernel/Atom.h
@@ -64,8 +64,8 @@ MANTID_KERNEL_DLL bool operator==(const Atom &left, const Atom &right);
 MANTID_KERNEL_DLL bool operator!=(const Atom &left, const Atom &right);
 /// Stream operator overload
 MANTID_KERNEL_DLL std::ostream &operator<<(std::ostream &out, const Atom &atom);
-MANTID_KERNEL_DLL Atom getAtom(const uint16_t z_number,
-                               const uint16_t a_number = 0);
+MANTID_KERNEL_DLL const Atom &getAtom(const uint16_t z_number,
+                                      const uint16_t a_number = 0);
 MANTID_KERNEL_DLL Atom getAtom(const std::string &symbol,
                                const uint16_t a_number = 0);