… Framework/Kernel/inc/MantidKernel/span.hpp