**Description of work.**

Using the proton number as an index leads to some substantial performance improvements.

I added two performance tests to show the difference

* test_z_performance: before: 0.246293s, after: 0.0214344s
* test_symbol_performance: before: 5.21779s, after: 0.310218s

Uses `tcb::span` suggested by @SimonHeybrock. Should that header be in a different place? 

**To test:**

<!-- Instructions for testing. -->

* Code review
* I'd suggest running the performance tests with and without changes to `Atom.h` and Atom.cpp`

*There is no associated issue.*

I added a line to the release notes, though I suspect this is more about code smells than something observed by users.

---

#### Reviewer ####

Please comment on the following ([full description](http://developer.mantidproject.org/IndividualTicketTesting/)):

##### Code Review #####

- Is the code of an acceptable quality?
- Does the code conform to the [coding standards](http://developer.mantidproject.org/Standards/)?
- Are the unit tests small and test the class in isolation?
- If there are changes in the release notes then do they describe the changes appropriately?

##### Functional Tests #####

- Do changes function as described? Add comments below that describe the tests performed?
- Do the changes handle unexpected situations, e.g. bad input?
- Has the relevant (user and developer) documentation been added/updated?

Does everything look good? Mark the review as **Approve**. A member of `@mantidproject/gatekeepers` will take care of it.