@peterfpeterson 2019-05-20
Should `span.hpp` be grabbed through an external project?