@quantumsteve 2019-05-20
@peterfpeterson I moved span to an external project, which reduces the size of this PR.