@quantumsteve 2019-05-20
`DOWNLOAD_NO_EXTRACT` wasn't available until CMake v3.6
https://github.com/Kitware/CMake/commit/1ba9b535478a2a73100a6b0e9bf01983b4774cb8